Moved all function declarations and Constants from GPhoCS.c & patch.c to GPhoCS.h & patch.h. Added CombStats & CombPrinter files and compiled them.
