+++ src/CombPrinter.c
@@ -5,58 +5,56 @@
  *	- returns 0
  ***********************************************************************************/
 
-#include "patch.c"
+#include "CombPrinter.h"
+#include "CombStats.h"
 
-int printCombStats (int iteration, FILE* file)		{
+#include "MCMCcontrol.h"
 
-	double logPrior = 0;
+#include "GPhoCS.h"
+#include "patch.h"
 
-	// print header
+
+#include <stdio.h>
+
+void printCombStats(int iteration, FILE* file){
+
+	return;
+
 	if(iteration<0) {
-		fprintf(file, "iter\tcoalStatFlat\tnumCoalFlat\tmigStatFlat\tnumMigFlat\tlogPrior\tlogLikelihood\tdataLogLikelihood\tgenealogyLogLikelihood");
-		fprintf(file, "\n");
+		printCombStatsHeader(file);
 
-		return 0;
-	}// end of if(iteration<0)
+		return;
+	}
 
-	logPrior = getLogPrior();
-
-
-	fprintf(file, "%7d", iteration);
-	fprintf(file, "\t%9f\t%9d\t%9f\t%9d\t%9f\t%9f\t%9f\t%9f",
-			genetree_stats_flat.coal_stats_flat,
-			genetree_stats_flat.num_coals_total,
-			genetree_stats_flat.mig_stats_flat,
-			genetree_stats_flat.num_migs_total,
-			logPrior,
-			dataState.logLikelihood*dataSetup.numLoci,
-			dataState.dataLogLikelihood,
-			dataState.genealogyLogLikelihood
-	);
-
+	fprintf(file, "%d\t", iteration);
+	for (int comb = 0; comb < dataSetup.popTree->numPops; comb++) {
+		printSpecificCombStats(comb, file);
+	}
+	for (int pop = 0; pop < dataSetup.popTree->numPops; pop++) {
+		printSpecificPopStats(pop, file);
+	}
+	for(int mig_band=0; mig_band<dataSetup.popTree->numMigBands; mig_band++) {
+		printSpecificMigBandStats(mig_band, file);
+	}
 	fprintf(file, "\n");
 	fflush(file);
-
-	return 0;
 }
-/** end of printCoalStats **/
 
-
-void printCladeStatsHeader(FILE* file){
+void printCombStatsHeader(FILE* file){
 	fprintf(file, "iteration\t");
 	for (int comb = 0; comb < dataSetup.popTree->numPops; comb++) {
-		printSpecificCladeHeader(comb, file);
+		printSpecificCombHeader(comb, file);
 	}
 	for (int pop = 0; pop < dataSetup.popTree->numPops; pop++) {
 		printSpecificPopHeader(pop, file);
 	}
 	for(int mig_band=0; mig_band<dataSetup.popTree->numMigBands; mig_band++) {
-			printSpecificMigHeader(mig_band, file);
+		printSpecificMigHeader(mig_band, file);
 	}
 
 	fprintf(file, "\n");
 }
-void printSpecificCladeHeader(int comb, FILE* file){
+void printSpecificCombHeader(int comb, FILE* file){
 	char* combName = dataSetup.popTree->popArray[comb].name;
 	fprintf(file, "%s_%s\t%s_%s\t",
 			combName, "coal_stats_total",
@@ -76,22 +74,8 @@ void printSpecificMigHeader(int mig_band, FILE* file){
 			migBandName, "_num_migs");
 }
 
-void printCladeStats(int iteration, FILE* file){
-	fprintf(file, "%d\t", iteration);
-	for (int comb = 0; comb < dataSetup.popTree->numPops; comb++) {
-		printSpecificCladeStats(comb, file);
-	}
-	for (int pop = 0; pop < dataSetup.popTree->numPops; pop++) {
-		printSpecificPopStats(pop, file);
-	}
-	for(int mig_band=0; mig_band<dataSetup.popTree->numMigBands; mig_band++) {
-		printSpecificMigBandStats(mig_band, file);
-	}
-	fprintf(file, "\n");
-	fflush(file);
-}
 
-void printSpecificCladeStats(int comb, FILE* file){
+void printSpecificCombStats(int comb, FILE* file){
 	fprintf(file, "%0.35f\t%d\t",
 			comb_stats[comb].coal_stats_total,
 			comb_stats[comb].num_coals_total);
