+++ src/CombPrinter.h
@@ -1,20 +1,18 @@
-#include <stdlib.h>
+#include <stdio.h>
 
 
 #ifndef SRC_COMBSTATSHEADER_H_
 #define SRC_COMBSTATSHEADER_H_
 
-/** clade stats printing functions **/
 
-void printCladeStats(int iteration, FILE* file);
+void printCombStats(int iteration, FILE* file);
 
-
-void printCladeStatsHeader(FILE* file);
-void printSpecificCladeHeader(int clade, FILE* file);
+void printCombStatsHeader(FILE* file);
+void printSpecificCombHeader(int clade, FILE* file);
 void printSpecificPopHeader(int pop, FILE* file);
 void printSpecificMigHeader(int mig_band, FILE* file);
-void printCladeStats(int iteration, FILE* file);
-void printSpecificCladeStats(int clade, FILE* file);
+void printCombStats(int iteration, FILE* file);
+void printSpecificCombStats(int clade, FILE* file);
 void printSpecificPopStats(int pop, FILE* file);
 void printSpecificMigBandStats(int mig_band, FILE* file);
